Implement new UI (#132)
* Squashed commit of the following:

commit ea91e1d9d8414e91f5e65fe69f9b3ac9eae61265
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Tue Aug 10 04:07:53 2021 -0400

    Refactor delete item using restx (#126)

    * refactor delete item

    * patch refactor_delete_item

    * touch up comments

commit b139f576276b790d30c2d4f0a9c3c71297197e79
Merge: d2a7e6f 03778ab
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Tue Aug 10 01:36:49 2021 -0400

    Merge pull request #125 from DevOps-Agile-Squad/refactor-get-item

    refactor get item and create an item

commit 03778aba258548b38c4dadfdd489cb96b552e6e0
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:53:28 2021 -0400

    adjust error string

commit b575ed6753614f43033f40ffc8085755624bf82d
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:44:03 2021 -0400

    fix bug: change 'id' in item_model into 'item_id' in routes.py

commit 0905b67f34b7906af35e78e4095691be21bcbadb
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:23:30 2021 -0400

    refactor-add-item

commit b2a9c2e0bb6edbc3821da8509e70c9179ba30bba
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 02:45:54 2021 -0400

    adjust error string

commit 3b2c0352e8aeda2d77850c78094939b721e11e27
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 02:37:20 2021 -0400

    replace raise with abort

commit b47179c5bc6ddb69aa41a57ed3211921c53f3eff
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 01:14:04 2021 -0400

    refactor get item

commit d2a7e6f2c66f2b624bbf850cdd71e8e3462f3723
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Mon Aug 9 00:05:26 2021 -0400

    introduce item model restx (#124)

commit 8ba02bcb2d5f0fae1fe18d8bea3babf272fdfff4
Merge: a132545 3026f90
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Sun Aug 8 23:25:14 2021 -0400

    Merge pull request #123 from DevOps-Agile-Squad/fix-swagger-fetch-error

    Fix swagger.json Fetch error

commit 3026f907e84bb8454d8bf933dba96d42aa2db227
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 23:14:44 2021 -0400

    change multiple locations into single location

commit a13254547e18fce552f87974a1129b70dc17c788
Merge: 770279a 148b486
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Sun Aug 8 20:48:21 2021 -0400

    Merge pull request #117 from DevOps-Agile-Squad/refactor-list

    Refactor list

commit 148b4868f9180bf252d6df9fb5897c446db0f855
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:41:54 2021 -0400

    add args as an additonal location where you can pull arguments down from requests

commit 73d11ab9f4df928178ee4f5a3100a6fc931034f9
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:19:43 2021 -0400

    delete unneseccery logging

commit 48f7b4ca7a7da4d062ba3e245debe420036760d4
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:18:35 2021 -0400

    fix bug: get 400 bad request when parse request

commit b25b544a54e5d6f03a4ff0869a8250a6b052aa20
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:08:06 2021 -0400

    refactor list

commit 770279a22d6ee927032086cebf8494973a3a5df4
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Sun Aug 8 09:01:15 2021 -0400

    commit used to trigger and test delivery pipeline (#116)

* solves `ElementClickInterceptedException`

* update jquery to v3.6.0

* change bootstrap files

* Squashed commit of the following:

commit 391d63fdd8de7cf4f9cb7c4235a29ba740c2f0b1
Merge: ea91e1d 9fd602c
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Tue Aug 10 21:40:37 2021 -0400

    Merge pull request #130 from DevOps-Agile-Squad/fix-500-type-error

    Fix 500 type error

commit 9fd602c6648f546277eab3f7ae0b73d01eba0543
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Tue Aug 10 19:11:34 2021 -0400

    specify the data type of parameters in a path (should be int) and modify the corresponding step method to send int but not string

commit dbbe16320727e48f2914f3d1a7a9a8270df19748
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Tue Aug 10 18:34:47 2021 -0400

    add validate=True

commit ea91e1d9d8414e91f5e65fe69f9b3ac9eae61265
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Tue Aug 10 04:07:53 2021 -0400

    Refactor delete item using restx (#126)

    * refactor delete item

    * patch refactor_delete_item

    * touch up comments

commit b139f576276b790d30c2d4f0a9c3c71297197e79
Merge: d2a7e6f 03778ab
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Tue Aug 10 01:36:49 2021 -0400

    Merge pull request #125 from DevOps-Agile-Squad/refactor-get-item

    refactor get item and create an item

commit 03778aba258548b38c4dadfdd489cb96b552e6e0
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:53:28 2021 -0400

    adjust error string

commit b575ed6753614f43033f40ffc8085755624bf82d
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:44:03 2021 -0400

    fix bug: change 'id' in item_model into 'item_id' in routes.py

commit 0905b67f34b7906af35e78e4095691be21bcbadb
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 23:23:30 2021 -0400

    refactor-add-item

commit b2a9c2e0bb6edbc3821da8509e70c9179ba30bba
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 02:45:54 2021 -0400

    adjust error string

commit 3b2c0352e8aeda2d77850c78094939b721e11e27
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 02:37:20 2021 -0400

    replace raise with abort

commit b47179c5bc6ddb69aa41a57ed3211921c53f3eff
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Mon Aug 9 01:14:04 2021 -0400

    refactor get item

commit d2a7e6f2c66f2b624bbf850cdd71e8e3462f3723
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Mon Aug 9 00:05:26 2021 -0400

    introduce item model restx (#124)

commit 8ba02bcb2d5f0fae1fe18d8bea3babf272fdfff4
Merge: a132545 3026f90
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Sun Aug 8 23:25:14 2021 -0400

    Merge pull request #123 from DevOps-Agile-Squad/fix-swagger-fetch-error

    Fix swagger.json Fetch error

commit 3026f907e84bb8454d8bf933dba96d42aa2db227
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 23:14:44 2021 -0400

    change multiple locations into single location

commit a13254547e18fce552f87974a1129b70dc17c788
Merge: 770279a 148b486
Author: JoanneWooW <85208330+JoanneWooW@users.noreply.github.com>
Date:   Sun Aug 8 20:48:21 2021 -0400

    Merge pull request #117 from DevOps-Agile-Squad/refactor-list

    Refactor list

commit 148b4868f9180bf252d6df9fb5897c446db0f855
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:41:54 2021 -0400

    add args as an additonal location where you can pull arguments down from requests

commit 73d11ab9f4df928178ee4f5a3100a6fc931034f9
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:19:43 2021 -0400

    delete unneseccery logging

commit 48f7b4ca7a7da4d062ba3e245debe420036760d4
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:18:35 2021 -0400

    fix bug: get 400 bad request when parse request

commit b25b544a54e5d6f03a4ff0869a8250a6b052aa20
Author: JoanneWooW <mw4708@nyu.edu>
Date:   Sun Aug 8 19:08:06 2021 -0400

    refactor list

commit 770279a22d6ee927032086cebf8494973a3a5df4
Author: pkutlesa <41484328+pkutlesa@users.noreply.github.com>
Date:   Sun Aug 8 09:01:15 2021 -0400

    commit used to trigger and test delivery pipeline (#116)

* fix error string format

* Improve logic for cancelling order

* change expected error code from 400 to 409

* remove old css files

* Add `Query orders by Item` to test UI change

* Change `Price` field to float

* remove unused error_handlers

* Design new Web UI

* Add new step to click radio option

+++ service/models.py
@@ -34,7 +34,7 @@ Item - An item object represents the product in an order.
     -----------
     order_id (fk integer) - the order number that the item is associated with
     quantity (integer) - the quantity of this item in the order
-    price (integer) - the price of the product
+    price (float) - the price of the product
     item_name (integer) - the name of the product
 """
 import logging
@@ -80,7 +80,7 @@ class Item(db.Model):
         'customer_order.id'), nullable=False)
     # If null quantity will be treated as 1.
     quantity = db.Column(db.Integer, nullable=True)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     # e.g., ball, balloon, etc.
     item_name = db.Column(db.String(120), nullable=False)
